+++ lib/cannery_web/components/invite_card.ex
@@ -42,6 +42,11 @@ defmodule CanneryWeb.Components.InviteCard do
         </h2>
       <% end %>
 
+      <.qr_code
+        content={Routes.user_registration_url(Endpoint, :new, invite: @invite.token)}
+        filename={@invite.name}
+      />
+
       <div class="flex flex-row flex-wrap justify-center items-center">
         <code
           id={"code-#{@invite.id}"}