+++ lib/cannery/accounts.ex
@@ -384,7 +384,7 @@ defmodule Cannery.Accounts do
   """
   @spec allow_registration?() :: boolean()
   def allow_registration? do
-    Application.get_env(:Cannery, CanneryWeb.Endpoint)[:registration] == "public" or
+    Application.get_env(:cannery, Cannery.Accounts)[:registration] == "public" or
       list_users_by_role(:admin) |> Enum.empty?()
   end