+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.8.1
 - Update dependencies
 - Show topbar on form submit/page refresh
+- Make loading/reconnection less intrusive
 
 # v0.8.0
 - Add search to catalog, ammo, container, tag and range index pages