+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update dependencies
 - Show topbar on form submit/page refresh
 - Make loading/reconnection less intrusive
+- Add QR code for invite link
 
 # v0.8.0
 - Add search to catalog, ammo, container, tag and range index pages